+++ History.md
@@ -1,6 +1,6 @@
 1.0.19 / 2015-??-??
 ==================
-  * Internal: Upgrade to Jasmine 2.3.0
+  * Internal: Upgrade to Jasmine 2.3.2
 
 1.0.18 / 2015-04-18
 ==================